+++ libs/base/src/sys/mem/slice/mod.rs
@@ -1,0 +1,12 @@
+// devela_base::sys::mem::slice
+//
+//! Slices of memory.
+//
+
+mod namespace; // Slice
+
+crate::structural_mods! { // _mods
+    _mods {
+        pub use super::namespace::Slice;
+    }
+}