+++ libs/base/src/sys/mem/slice/namespace.rs
@@ -1,4 +1,4 @@
-// devela::sys::mem::slice::namespace
+// devela_base::sys::mem::slice::namespace
 //
 //! Defines the [`Slice`] namespace.
 //
@@ -445,7 +445,7 @@ impl<T> Slice<T> {
     ///
     /// # Example
     /// ```
-    /// # use devela::Slice;
+    /// # use devela_base::Slice;
     /// let v = [1, 2, 3, 4, 5, 6];
     /// assert_eq!(Slice::lsplit(&v, 3), &[1, 2, 3]);
     /// assert_eq!(Slice::lsplit(&v, 0), &[] as &[i32]);
@@ -463,7 +463,7 @@ impl<T> Slice<T> {
     ///
     /// # Example
     /// ```
-    /// # use devela::Slice;
+    /// # use devela_base::Slice;
     /// let mut v = [1, 2, 3, 4, 5, 6];
     /// assert_eq!(Slice::lsplit_mut(&mut v, 3), &mut [1, 2, 3]);
     /// assert_eq!(Slice::lsplit_mut(&mut v, 0), &mut [] as &mut [i32]);
@@ -484,7 +484,7 @@ impl<T> Slice<T> {
     ///
     /// # Example
     /// ```
-    /// # use devela::Slice;
+    /// # use devela_base::Slice;
     /// let v = [1, 2, 3, 4, 5, 6];
     /// assert_eq!(Slice::rsplit(&v, 3), &[4, 5, 6]);
     /// assert_eq!(Slice::rsplit(&v, 0), &[] as &[i32]);
@@ -503,7 +503,7 @@ impl<T> Slice<T> {
     ///
     /// # Example
     /// ```
-    /// # use devela::Slice;
+    /// # use devela_base::Slice;
     /// let mut v = [1, 2, 3, 4, 5, 6];
     /// assert_eq!(Slice::rsplit_mut(&mut v, 3), &mut [4, 5, 6]);
     /// assert_eq!(Slice::rsplit_mut(&mut v, 0), &mut [] as &mut [i32]);
@@ -528,7 +528,7 @@ impl<T> Slice<T> {
     ///
     /// # Example
     /// ```
-    /// # use devela::Slice;
+    /// # use devela_base::Slice;
     /// let v = [1, 2, 3, 4, 5, 6];
     /// assert_eq!(Slice::msplit_left(&v, 0), &[] as &[i32]);
     /// assert_eq!(Slice::msplit_left(&v, 1), &[3]);
@@ -560,7 +560,7 @@ impl<T> Slice<T> {
     ///
     /// # Example
     /// ```
-    /// # use devela::Slice;
+    /// # use devela_base::Slice;
     /// let mut v = [1, 2, 3, 4, 5, 6];
     /// assert_eq!(Slice::msplit_left_mut(&mut v, 0), &mut [] as &mut [i32]);
     /// assert_eq!(Slice::msplit_left_mut(&mut v, 1), &mut [3]);
@@ -593,7 +593,7 @@ impl<T> Slice<T> {
     ///
     /// # Example
     /// ```
-    /// # use devela::Slice;
+    /// # use devela_base::Slice;
     /// let v = [1, 2, 3, 4, 5, 6];
     /// assert_eq!(Slice::msplit_right(&v, 0), &[] as &[i32]);
     /// assert_eq!(Slice::msplit_right(&v, 1), &[4]);
@@ -624,7 +624,7 @@ impl<T> Slice<T> {
     /// If `len > slice.len()` returns the full `slice`.
     /// # Example
     /// ```
-    /// # use devela::Slice;
+    /// # use devela_base::Slice;
     /// let mut v = [1, 2, 3, 4, 5, 6];
     /// assert_eq!(Slice::msplit_right_mut(&mut v, 0), &mut [] as &mut[i32]);
     /// assert_eq!(Slice::msplit_right_mut(&mut v, 1), &mut [4]);